traj: allow specifying the trajectory, observation and time dimensions/variables
Not 100% sure if this is necessary. I was trying to avoid the problem
where there's another dimension to time (say 'campaign'), but that's
another issue. This is mostly to allow users to specify the dimensions
manually when there is ambiguity. Maybe it would be better to allow
specifying the longitude and latitude variables (e.g. in the case when
there are interpolated values for the IMU, sometimes you want one,
sometimes the other).